Created EntityBasePage which can be inherited by all table pages without clickNewFolder() (like MyAssignmentsPage). EntityBaseTablePage inherits everything from EntityBasePage and adds clickNewFolder(). Corrected  listButton's xpath
